backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator",
    "fastapi",
    "pydantic",
    "pydantic-settings",
    "pyyaml",
    "requests",
    "uvicorn[standard]",
    "injector",
    "kubernetes",
    "aiosmtplib",
    "loguru"
]
